+++ src/Servers/OSX.php
@@ -22,7 +22,7 @@ final class OSX implements Implementation
 {
     private CpuFacade $cpu;
     private MemoryFacade $memory;
-    private Processes\Unix $processes;
+    private Processes $processes;
     private LoadAverageFacade $loadAverage;
     private Disk $disk;
 
@@ -30,7 +30,7 @@ final class OSX implements Implementation
     {
         $this->cpu = new CpuFacade($control->processes());
         $this->memory = new MemoryFacade($control->processes(), $path);
-        $this->processes = Processes\Unix::osx($clock, $control->processes());
+        $this->processes = Processes::osx($clock, $control->processes());
         $this->loadAverage = new LoadAverageFacade;
         $this->disk = Disk::of($control->processes());
     }